Add files via upload
Python script for taking screenshots by using the tkinter and pyautogui modules of python. The tkinter allows to create widgets like various buttons with the options of taking the new screenshot or quit it. The screen is captured using the "screenshot" function of the pyautogui module. Then  the image gets dispalyed on the screen with the help of "show" function.